Behave gracefully when unlinking/relinking twice

+++ atomica/model.py
@@ -651,6 +651,11 @@ class Model(object):
     def unlink(self):
         # Break cycles when deepcopying or pickling by swapping them for UIDs
         # Primary storage is in the comps, links, and outputs properties
+
+        # If we are already unlinked, do nothing
+        if self.pars_by_pop is None:
+            return
+
         for pop in self.pops:
             pop.unlink()
         if self.pset is not None:
@@ -659,6 +664,11 @@ class Model(object):
 
     def relink(self):
         # Need to enumerate objects at Model level because transitions link across pops
+
+        # If we are already linked, do nothing
+        if self.pars_by_pop is not None:
+            return
+
         objs = {}
         for pop in self.pops:
             for obj in pop.comps + pop.characs + pop.pars + pop.links:
